mikrobot-mpc/requirement.txt
kafka-python
python-dotenv

pip install kafka-python python-dotenv
pip freeze > requirements.txt
